优化对长按的检测

+++ Unity/Assets/Test/Scripts/TestEventDelivery.cs
@@ -17,7 +17,7 @@ public class TestEventDelivery : MonoBehaviour, IPointerClickHandler, IBeginDrag
         delivery.LongPress += CallBack;
     }
 
-    private void CallBack(GameObject go, float pressTime)
+    private void CallBack(GameObject go, double pressTime)
     {
         if (go != gameObject)
         {
